+++ src/embed/terminators/go.ps1
@@ -5,7 +5,7 @@ Write-Host ">>> Removing Golang via MSIExec..."
 $line = (& go.exe version) 
 
 $version = ($line | Select-String "go\d+\.\d+\.\d+").Matches.Value
-$arch    = ($line | Select-String "(?<=/)[^ ]+$").Matches.Value
+$arch = ($line | Select-String "(?<=/)[^ ]+$").Matches.Value
 
 & msiexec.exe /x "go$version.windows-$arch.msi" /q