+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Dates are in the DD-MM-YYYY format.
 
+## Unreleased
+
+### Changed
+
+- Now the `terminate` command properly removes PATH entries for Deno and Bun; and it is also better at removing Node.js from Windows.
+
 ## [5.2.3] (07-02-2026)
 
 ### Added